Test bishop movement This is the behaviour that is intended for the Bishop. The next step is to develop this behaviour
